backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.32
pymysql==1.1.1
python-dotenv==1.0.1
pydantic==2.8.2

